refactor(aviation): extract location data + predicates into location.rs
